code movement, delete unused code
